Added drag window. Added metrics module
