+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Debugging/FPSDisplay/DebugValue.cs
@@ -22,6 +22,6 @@ public enum DebugValueEnum
     Skybox_Config, Skybox_Duration, Skybox_GameTime, Skybox_UTCTime, General_Network, General_Realm, General_NearbyPlayers, FPS, FPS_HiccupsInTheLast1000, FPS_HiccupsLoss, FPS_BadFramesPercentiles,
     Scene_ProcessedMessages, Scene_PendingOnQueue, Scene_Poly, Scene_Textures, Scene_Materials, Scene_Entities, Scene_Meshes, Scene_Bodies, Scene_Components, Scene_Name,
     Memory_Used_JS_Heap_Size, Memory_Limit_JS_Heap_Size, Memory_Total_JS_Heap_Size,
-    Memory_Total_Allocated, Memory_Reserved_Ram, Memory_Mono
+    Memory_Total_Allocated, Memory_Reserved_Ram, Memory_Mono, Memory_Graphics_Card
 }
