+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Debugging/FPSDisplay/DragUIHandler.cs
@@ -1,12 +1,17 @@
 using UnityEngine;
 using UnityEngine.EventSystems;
- 
+using UnityEngine.UI;
+
 public class DragUIHandler : MonoBehaviour, IBeginDragHandler, IDragHandler, IEndDragHandler {
  
     private Vector2 pointerOffset;
     private RectTransform canvasRectTransform;
     private RectTransform panelRectTransform;
- 
+    private Vector3 startLocalScale;
+    private Vector2 originalPivot;
+    [SerializeField] private float scaleDownOnBeginDrag = 1f;
+    private bool jumpOneFrame;
+
     public void Start(){
        
         Canvas canvas = GetComponentInParent<Canvas> ();
@@ -14,12 +19,17 @@ public class DragUIHandler : MonoBehaviour, IBeginDragHandler, IDragHandler, IEn
             canvasRectTransform = canvas.transform as RectTransform;
             panelRectTransform = transform as RectTransform;
         }
+        startLocalScale = panelRectTransform.localScale;
+        originalPivot = panelRectTransform.pivot;
     }
  
     #region IBeginDragHandler implementation
  
     public void OnBeginDrag (PointerEventData eventData)
     {
+        RectTransformUtility.ScreenPointToLocalPointInRectangle (panelRectTransform, eventData.position, eventData.pressEventCamera, out Vector2 localPointerForNewPivot);
+        SetPivot(panelRectTransform, Rect.PointToNormalized(panelRectTransform.rect, localPointerForNewPivot));
+        panelRectTransform.localScale = startLocalScale * scaleDownOnBeginDrag;
         RectTransformUtility.ScreenPointToLocalPointInRectangle (panelRectTransform, eventData.position, eventData.pressEventCamera, out pointerOffset);
     }
  
@@ -32,8 +42,7 @@ public class DragUIHandler : MonoBehaviour, IBeginDragHandler, IDragHandler, IEn
         if(panelRectTransform == null){
             return;
         }
-        Vector2 localPointerPosition;
-        if(RectTransformUtility.ScreenPointToLocalPointInRectangle (canvasRectTransform, eventData.position, eventData.pressEventCamera, out localPointerPosition)){
+        if(RectTransformUtility.ScreenPointToLocalPointInRectangle (canvasRectTransform, eventData.position, eventData.pressEventCamera, out Vector2 localPointerPosition)){
             panelRectTransform.localPosition = localPointerPosition - pointerOffset;
         }
     }
@@ -44,8 +53,19 @@ public class DragUIHandler : MonoBehaviour, IBeginDragHandler, IDragHandler, IEn
  
     public void OnEndDrag (PointerEventData eventData)
     {
- 
+        panelRectTransform.localScale = startLocalScale;
+        SetPivot(panelRectTransform, originalPivot);
     }
  
     #endregion
+    
+    private void SetPivot(RectTransform target, Vector2 pivot)
+    {
+        if (!target) return;
+        var offset= pivot - target.pivot;
+        offset.Scale(target.rect.size);
+        var wordlPos= target.position + target.TransformVector(offset);
+        target.pivot = pivot;
+        target.position = wordlPos;
+    }
 }
