+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Debugging/FPSDisplay/MetricsModules/GeneralDebugMetricModule.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 526992cfb1178b045a15cddde0328309
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
