+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Debugging/FPSDisplay/MetricsModules/IDebugMetricModule.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface IDebugMetricModule : IDisposable
+{
+
+    public void SetUpModule(Dictionary<DebugValueEnum, Func<string>> updateValueDictionary);
+    public void UpdateModule();
+    public void EnableModule();
+    public void DisableModule();
+
+}
