+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Debugging/FPSDisplay/MetricsModules/MemoryJSDebugMetricModule.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 831f781895524b04f827d359251e8ebc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
