+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Debugging/FPSDisplay/MetricsModules/SceneDebugMetricModule.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae17f03a3e573cf47854de6ed59047cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
