+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Debugging/FPSDisplay/MetricsModules/SkyboxDebugMetricModule.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c72484e7fb9357a4d9c68c4d866efa4b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
