+++ unity-renderer/Assets/Textures/MoveIcon.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:59985a1c376121c6864b38f88cf14718226d10d5baf64402c79e7da5b0d47fe4
+size 492
